refactor: 패키지 정리를 한다
