+++ android/app/src/main/java/com/digginroom/digginroom/feature/scrap/ScrapBindingAdapter.kt
@@ -5,6 +5,7 @@ import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.digginroom.digginroom.R
+import com.digginroom.digginroom.feature.scrap.adapter.ScrapAdapter
 import com.digginroom.digginroom.model.RoomModel
 
 object ScrapBindingAdapter {
