+++ android/app/src/main/java/com/digginroom/digginroom/feature/scrap/adapter/ScrappedRoomViewHolder.kt
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.feature.scrap
+package com.digginroom.digginroom.feature.scrap.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
